Kodutoo6

+++ src/main/java/springalgus/Rakendus6.java
@@ -12,7 +12,7 @@ public class Rakendus6 {
     @RequestMapping("/kehamassiindeks")
     String korrutamine(int arv1, int arv2) {
         String vastus = "vastus";
-        int vaartus = (int) (arv1/(arv2*arv2));
+        int vaartus = arv1/(arv2*arv2);
         vastus = "Vastus on: " + vaartus;
             return vastus;
     }
